KOGITO-1436: setContent() refactoring
