+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-kogito-runtime/src/test/java/org/drools/workbench/screens/scenariosimulation/webapp/client/editor/ScenarioSimulationEditorKogitoRuntimeScreenTest.java
@@ -60,7 +60,6 @@ public class ScenarioSimulationEditorKogitoRuntimeScreenTest {
         scenarioSimulationEditorKogitoRuntimeScreenSpy = spy(new ScenarioSimulationEditorKogitoRuntimeScreen(placeManagerMock) {
             {
                 this.scenarioSimulationEditorKogitoWrapper = scenarioSimulationEditorKogitoWrapperMock;
-                this.scenarioKogitoCreationPopupPresenter = scenarioKogitoCreationPopupPresenterMock;
             }
         });
     }
